+++ pkg/alerting/alertmanager/error.go
@@ -2,6 +2,8 @@ package alertmanager
 
 import "fmt"
 
+// APIError represents an error response from the Mimir Alertmanager API,
+// capturing the HTTP status code and response body for diagnostics.
 type APIError struct {
 	Code    int
 	Message string